refactor(sync): split sync.py into modular components
Extract logically grouped functions from the monolithic sync.py (43.8 KB)
into focused modules to improve context efficiency for AI agents:

- ado_parser.py: ADOAssignedUser dataclass and get_task_user
- matching.py: matching_user for ADO/Asana user resolution
- task_factory.py: create_asana_task_body, _build_asana_task_body,
  _save_created_task, _retry_create_without_due_date, _CLOSED_STATES
- asana_client.py: Asana API lookup functions (workspace, project,
  tasks, tags, users) and _get_deactivated_user_gids

sync.py retains orchestration logic and re-exports all moved symbols
so the public API and all existing test patches are unchanged.

Also suppress pre-existing F401 in tests/sync/test_sync.py via
pyproject.toml per-file-ignores.
